Rename: 생년월일 선택기 위치 이동
