+++ src/app/my-center/members/panel/birth-date-picker/DepthToggle.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React from "react";
-import SenifitToggleButtonGroup from "../SenifitToggleButtonGroup";
+import SenifitToggleButtonGroup from "../../../../../components/SenifitToggleButtonGroup";
 import { Box, Typography } from "@mui/material";
 import useMedia from "@/hooks/useMedia";
 import { ISenifitToggleOption } from "@/types/IToggleButton";
