+++ src/app/my-center/members/panel/birth-date-picker/Title.tsx
@@ -1,6 +1,6 @@
 import { Box, Button, Stack, Typography } from "@mui/material";
 import React from "react";
-import { CrossIcon } from "../icons";
+import { CrossIcon } from "../../../../../components/icons";
 
 const Title = ({ closeDatePicker }: { closeDatePicker: () => void }) => {
   return (
